+++ Changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Added conversions from `Extent2D` to `Extent3D` and `Rect2D` (#557)
+
 ## [0.35.1] - 2022-01-18
 
 ### Added